# MAINT: speed up array binary operations