Matmul was very slow before!

```python3
import numpy as np
from marray import numpy as mnp
rng = np.random.default_rng(893458972872134)
shape = (1000, 1000)

data_a = rng.random(shape)
mask_a = rng.random(shape) > 0.5
data_b = rng.random(shape)
mask_b = rng.random(shape) > 0.5

xa = mnp.asarray(data_a.copy(), mask=mask_a.copy())
xb = mnp.asarray(data_b.copy(), mask=mask_b.copy())

ya = np.ma.masked_array(data_a.copy(), mask=mask_a.copy())
yb = np.ma.masked_array(data_b.copy(), mask=mask_b.copy())

# %timeit data_a @ data_b
# 8.2 ms ± 58.8 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# %timeit xa @ xb
# before: 1.5 s ± 44.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# after: 21.3 ms ± 1.36 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# %timeit ya @ yb
# 9.91 ms ± 241 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Still not as fast as NumPy masked array, but that might take a little more work. Let's start with this and improve further once gh-85 is ready.