@jorenham 2025-01-28
maybe `__slots__` could help speed it up a bit more 🤔